+++ core/admin.py
@@ -2,8 +2,6 @@
 
 from django.contrib import admin
 from .models import Document, Address
-from django.http import HttpResponse
-import csv
 # Register your models here.
 
 
@@ -16,23 +14,6 @@ class DocumentAdmin(admin.ModelAdmin):
 
     # reorganize fields
     fields = ('title', 'slug', 'content',)
-
-    actions = ["export_as_csv"]
-
-    def export_as_csv(self, request, queryset):
-        meta = self.model._meta
-        field_names = [field.name for field in meta.fields]
-
-        response = HttpResponse(content_type='text/csv')
-        response['Content-Disposition'] = 'attachment; filename={}.csv'.format(
-            meta)
-        writer = csv.writer(response)
-        writer.writerow(field_names)
-        for obj in queryset:
-            row = writer.writerow([getattr(obj, field)
-                                   for field in field_names])
-        return response
-    export_as_csv.short_description = "Exporter au format CSV"
 
 
 @admin.register(Address)
