+++ users/admin.py
@@ -7,13 +7,9 @@ from django.utils.translation import ugettext_lazy as _
 from visits.admin import ParticipationInline
 
 from .models import User
+from django.http import HttpResponse
+import csv
 
-actions = ["export_as_csv"]
-
-def export_as_csv(self, request, queryset):
-    pass
-
-export_as_csv.short_description = "Export Selected"
 
 class UserParticipationInline(ParticipationInline):
     """Inline for Participation on the User admin panel.
@@ -63,3 +59,20 @@ class CustomUserAdmin(UserAdmin):
     )
 
     inlines = (UserParticipationInline, )
+
+    actions = ["export_as_csv"]
+
+    def export_as_csv(self, request, queryset):
+        meta = self.model._meta
+        field_names = [field.name for field in meta.fields]
+
+        response = HttpResponse(content_type='text/csv')
+        response['Content-Disposition'] = 'attachment; filename={}.csv'.format(
+            meta)
+        writer = csv.writer(response)
+        writer.writerow(field_names)
+        for obj in queryset:
+            row = writer.writerow([getattr(obj, field)
+                                   for field in field_names])
+        return response
+    export_as_csv.short_description = "Exporter au format CSV"
